Fix bug/auto generate the threshold gates
The parent commit accidentally swapped A and B arguments and prompted
the creation of `mk_lut.rs` to auto generate the threshold gates from
a boolean function + argument names.

Took the opportunity to reverse the mapping of gate inputs to LUT
inputs as apparently "D" inputs are the fastest (and "A" slowest?).

Note, this change is currently UNTESTED on hardware.